CHOP-74. Move Kiji dependencies to provided scope.
This commit modifies the build so that Kiji dependencies are in
provided scope. It also modifies the express script to expect
KIJI_HOME and KIJI_MR_HOME to be set so their dependencies can be
added to the classpath and tmpjars. Some updates on Kiji dependency
versions are also made. Additionally, RELEASE_NOTES.txt is added
to the project distribution, and a DDL file now only used in tests is
removed.

Reviewed by: robertchu