EXP-55. Allow KijiExpress to build against CDH4.1.2 and CDH4.2.1
This commit modifies the pom to include a profile that can build against
CDH4.1.2 (the default) and against CDH4.2.1. The profile that can build
against CDH4.2.1 also builds against compatible versions of kiji-schema
and kiji-mapreduce. This commit also includes changes to disable schema
validation in KijiSchema 1.1.0, as currently the validation is too strict
to be compatible with KijiExpress.

Reviewed by: robertchu