EXP-85. packAvro now expects to pack Scala types, not Java types.
The method packAvro in KijiPipe is used to pack multiple Scalding
tuple fields into one AvroRecord. To do this, the values in the tuple
fields must be converted to instances of AvroValue (really AvroValue's
ancestors). Before this commit, KijiPipe used AvroUtil#wrapGenericAvro
to perform this conversion. This is incorrect, however, as
AvroUtil#wrapGenericAvro expects Java types as input, but the tuple
field values to be packed should be Scala types. This commit changes
packAvro to use AvroUtil#scalaToGenericAvro instead, which expects
Scala types to convert to AvroValues.

Reviewed by: juliet