EXP-87. Fixup on entity id changes.
This commit updates the RELEASE_NOTES to include information about the
recent entity id change. It also fixes a bug in the method putTuple of
KijiScheme, which prevented the usage of HashedEntityId. putTuple has
been made general to all kinds of entity ids. This fixes a problem in
the express examples that caused test failures.

Reviewed by: daisy