EXP-16. Add support for writing to HFiles.
Add support for express/scalding to write to HFiles. Basically a client would
have to subclass HFileKijiJob and specify (among other things) the "output" flag
pointing to the Kiji table eventually being bulk loaded into and the "hFileOutput"
flag referring to where the HFiles will live. The HFileKijiJob has some methods
overridden that will setup the right output formats and if necessary kick off a
secondary scalding job that will perform a map-only + custom identity reducer
using total order partitioning to format the HFile KeyValues into a set of
HFiles that can be bulk loaded into HBase. The reason for this is because a
Cascading flow may already have a reducer + partitioner attached and modifying
this seemed dangerous; however launching a secondary job that we know the
properties of is much safer.

Reviewer: daisy