EXP-301. Fix the issue of writing to HFiles belonging to layout 1.3 tables.
When writing to an HFile whose columns are validated in DEVELOPER or STRICT
mode, Schema throws an exception because the cell spec has no column URI. This
was fixed by using a different encoder wrapper that sets the URI of the cell
spec. Along the way, also fixed a problem in how source/sink fields are set
which prevents duplicate field exceptions in Cascading.

Reviewer: robert